requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.1
django-ckeditor==5.9.0
django-el-pagination==3.3.0
django-embed-video==1.3.3
django-js-asset==1.2.2
django-leaflet==0.27.1
django-location-field==2.1.0
django-nested-inline==0.4.2
django-smart-selects==1.5.9
djangorestframework==3.11.1
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.5
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.10
webencodings==0.5.1
